add autoload-dev only to travis config
