+++ composer.json
@@ -18,11 +18,6 @@
             "FormBuilderBundle\\": "src/FormBuilderBundle"
         }
     },
-    "autoload-dev": {
-        "psr-4": {
-            "DachcomBundle\\Test\\FormBuilder\\Services\\": "tests/_support/FormBuilder/Services"
-        }
-    },
     "extra": {
         "pimcore": {
             "bundles": [
