+++ tests/_etc/config/bundle/symfony/config_default.yml
@@ -1,11 +1,11 @@
 services:
-    DachcomBundle\Test\FormBuilder\Services\TestSimpleDynamicChoices:
+    DachcomBundle\Test\App\Services\TestSimpleDynamicChoices:
         autowire: true
         public: false
         tags:
             - { name: form_builder.dynamic_choice_builder, label: 'Simple Test Selector' }
 
-    DachcomBundle\Test\FormBuilder\Services\TestAdvancedDynamicChoices:
+    DachcomBundle\Test\App\Services\TestAdvancedDynamicChoices:
         autowire: true
         public: false
         tags:
