+++ tests/_support/App/Services/TestAdvancedDynamicChoices.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace DachcomBundle\Test\FormBuilder\Services;
+namespace DachcomBundle\Test\App\Services;
 
-use FormBuilderBundle\Form\ChoiceBuilderInterface;
+use FormBuilderBundle\Form\AdvancedChoiceBuilderInterface;
 use Symfony\Component\Form\CallbackTransformer;
 use Symfony\Component\Form\FormBuilderInterface;
 
-class TestSimpleDynamicChoices implements ChoiceBuilderInterface
+class TestAdvancedDynamicChoices implements AdvancedChoiceBuilderInterface
 {
     protected $builder;
 
@@ -47,6 +47,31 @@ class TestSimpleDynamicChoices implements ChoiceBuilderInterface
                 }
             }
         ));
+    }
+
+    public function getChoiceValue($value = null)
+    {
+        return $value . '-custom-value';
+    }
+
+    public function getChoiceLabel($element, $key, $index)
+    {
+        return $key . ' Custom Label';
+    }
+
+    public function getChoiceAttributes($element, $key, $index)
+    {
+        return ['class' => 'special-choice-class'];
+    }
+
+    public function getGroupBy($element, $key, $index)
+    {
+        return 'Group A';
+    }
+
+    public function getPreferredChoices($element, $key, $index)
+    {
+        return $key === 'Entity 5';
     }
 
     public function getList()
