+++ tests/_support/App/Services/TestSimpleDynamicChoices.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace DachcomBundle\Test\FormBuilder\Services;
+namespace DachcomBundle\Test\App\Services;
 
-use FormBuilderBundle\Form\AdvancedChoiceBuilderInterface;
+use FormBuilderBundle\Form\ChoiceBuilderInterface;
 use Symfony\Component\Form\CallbackTransformer;
 use Symfony\Component\Form\FormBuilderInterface;
 
-class TestAdvancedDynamicChoices implements AdvancedChoiceBuilderInterface
+class TestSimpleDynamicChoices implements ChoiceBuilderInterface
 {
     protected $builder;
 
@@ -47,31 +47,6 @@ class TestAdvancedDynamicChoices implements AdvancedChoiceBuilderInterface
                 }
             }
         ));
-    }
-
-    public function getChoiceValue($value = null)
-    {
-        return $value . '-custom-value';
-    }
-
-    public function getChoiceLabel($element, $key, $index)
-    {
-        return $key . ' Custom Label';
-    }
-
-    public function getChoiceAttributes($element, $key, $index)
-    {
-        return ['class' => 'special-choice-class'];
-    }
-
-    public function getGroupBy($element, $key, $index)
-    {
-        return 'Group A';
-    }
-
-    public function getPreferredChoices($element, $key, $index)
-    {
-        return $key === 'Entity 5';
     }
 
     public function getList()
